refactor: simplify K1
Reuse: the MCP source walker and the tool-registration regex existed in two
copies (tool-count-drift, license-gate-coverage); one binding now lives in
tests/helpers/mcp-source.ts. The E2E pull-path spec's three near-identical
`callTool` narrows collapse onto one `callData<T>` over the shared
`ToolResponse` type. The coverage-gate specifier resolver folds its three
`.js`/extensionless branches into one.

Simplification: the axe `incomplete` ceiling's rationale moved off the
annotation-card surface entry (where it read as card-specific) onto
`INCOMPLETE_BASELINE`, which is what it actually governs; duplicated
"what this does not buy" prose dropped from axe-incomplete.ts. Unused
`InboxData.mode` removed.

Efficiency: tool-count-drift rescanned the whole MCP tree's registrations
three times; hoisted to one REGISTERED/TOTAL/ACTIVE binding.
