+++ tests/e2e/accessibility.spec.ts
@@ -324,29 +324,6 @@ const SURFACES: Surface[] = [
     // `color-contrast` is not in `disableRules`, so the rule DOES run — this is
     // a bucketing gap, not a disabled rule.
     //
-    // Since #1721 the loop below also pins a per-surface CEILING on
-    // color-contrast `incomplete` nodes (`tests/e2e/axe-incomplete-baseline.json`,
-    // option 3 from the issue). Read what that does and does not buy, because
-    // the obvious reading is again wrong: it makes a JUMP loud — a change that
-    // pushes more nodes into the bucket reds this suite — and it forces a newly
-    // added surface to be measured rather than silently unmeasured. It does NOT
-    // say any individual incomplete node is acceptable, and it is NOT a contrast
-    // gate for the user tint.
-    //
-    // RE-MEASURED 2026-09-06, and the result is not what the paragraph above
-    // predicts — record it as data, not as a resolution. Repainting
-    // `--tandem-author-user-bg` to `#4a4a4a` and running `-g "annotation card"`
-    // on the reserved ports: LIGHT and WARM now FAIL on `violations` (axe 4.13
-    // reports contrast 1.21 and 1.47 against `#4a4a4a`), while DARK stays green
-    // — and in all three themes the color-contrast `incomplete` count stayed at
-    // 4, its baseline. So the new pin did NOT catch this mutation anywhere; the
-    // pre-existing `violations` assertion caught two thirds of it, which the
-    // filed measurement says it did not. Whether that is an axe/Chromium
-    // difference or a change in the tints since is exactly the unestablished
-    // mechanism below. It bounds what option 3 buys: a jump detector, not a
-    // contrast gate — and it does not narrow what options (1) and (2) on #1721
-    // still have to explain.
-    //
     // The exact mechanism is NOT established, and an earlier version of this
     // comment asserted one it could not support. That run observed 25
     // color-contrast nodes under `incomplete` against 17 under `passes`, but a
@@ -501,9 +478,28 @@ const SURFACES: Surface[] = [
 ];
 
 /**
- * The committed per-surface ceiling (#1721). Read once, at module scope, with
- * `readFileSync` rather than a JSON import — the file's existing style, and no
- * `resolveJsonModule` dependency.
+ * The committed per-surface ceiling on axe's color-contrast `incomplete` nodes
+ * (#1721, option 3). It applies to EVERY surface above, not just the annotation
+ * card whose comment records why the bucket matters.
+ *
+ * What it buys, because the obvious reading is wrong: it makes a JUMP loud — a
+ * change that pushes more nodes into the bucket reds this suite — and it forces
+ * a newly added surface to be measured rather than silently unmeasured. It does
+ * NOT say any individual incomplete node is acceptable, and it is NOT a contrast
+ * gate.
+ *
+ * RE-MEASURED 2026-09-06 — data, not a resolution. Repainting
+ * `--tandem-author-user-bg` to `#4a4a4a` and running `-g "annotation card"`:
+ * LIGHT and WARM FAIL on `violations` (axe 4.13 reports contrast 1.21 and 1.47),
+ * DARK stays green, and the `incomplete` count stayed at its baseline of 4 in
+ * all three. So this pin caught that mutation nowhere and the pre-existing
+ * `violations` assertion caught two thirds of it — which the filed measurement
+ * says it did not. Whether that is an axe/Chromium difference or a tint change
+ * since is the unestablished mechanism recorded on the annotation-card surface.
+ * It does not narrow what options (1) and (2) on #1721 still have to explain.
+ *
+ * Read once with `readFileSync` rather than a JSON import — the file's existing
+ * style, and no `resolveJsonModule` dependency.
  */
 const INCOMPLETE_BASELINE = JSON.parse(
   readFileSync(new URL("./axe-incomplete-baseline.json", import.meta.url), "utf8"),
@@ -537,9 +533,7 @@ for (const theme of ["light", "dark", "warm"] as const) {
           // expressions is how a one-word slip to `…(results.violations)` ships a
           // permanently-green gate: violations is asserted empty on the next
           // line, so every key would seed as 0 and `observed <= baseline` would
-          // always hold. With one binding, a mis-wired gate can only produce an
-          // all-zero committed baseline — visible in the file, not hidden in a
-          // passing run. The log is deliberately permanent and ungated so a
+          // always hold. The log is deliberately permanent and ungated so a
           // re-seed can read CI's numbers without a code change.
           const key = `${theme} / ${surface.name}`;
           const incompleteNodes = colorContrastNodeCount(results.incomplete);
