+++ tests/e2e/axe-incomplete.ts
@@ -1,15 +1,8 @@
 /**
- * A per-surface ceiling on axe's `color-contrast` **incomplete** nodes (#1721).
- *
- * `accessibility.spec.ts` asserted `results.violations` only. axe has three
- * buckets, and a contrast failure that lands in `incomplete` was therefore
- * unasserted by construction — measured: a mid-grey annotation card behind dark
- * body text reported `violations: []` in all three themes.
- *
- * This is option 3 from the issue: a count pin. It makes a JUMP loud. It does
- * NOT say any individual incomplete node is acceptable, and it is not a contrast
- * gate — deciding what these nodes mean is what keeps #1721 open, and the gate
- * deliberately does not depend on that answer.
+ * The pure half of #1721's per-surface ceiling on axe's `color-contrast`
+ * **incomplete** nodes. Why the ceiling exists, what it measured, and what it
+ * does NOT buy are on `INCOMPLETE_BASELINE` in `accessibility.spec.ts`; kept
+ * there so the history has one home.
  *
  * Not a `.spec.ts`/`.test.ts`: Playwright's default `testMatch` under
  * `testDir: "tests/e2e"` must not collect it, and neither must vitest's node
