+++ tests/e2e/keyboard-a11y.spec.ts
@@ -263,16 +263,13 @@ test("the rail's panel switcher exposes which panel is selected", async ({ page 
   const annotations = page.locator("[data-testid='annotations-tab']");
   const chat = page.locator("[data-testid='chat-tab']");
 
-  // This carried a layout-mode guard — `if ((await annotations.count()) === 0)
-  // test.skip(true, …)` — for the three-panel layout, which renders both panels
-  // side by side with no tab buttons at all (see `switchToAnnotationsTab` in
-  // helpers.ts). MEASURED 2026-09-06 with `npx playwright test
-  // keyboard-a11y.spec.ts -g "panel switcher"` on the reserved ports: the test
-  // reported `passed`, not `skipped`, so the guard did not fire and the
-  // assertions below ran. It was dead weight of the #1529 shape — a runtime
-  // skip that nothing exercises reads exactly like a pass — so the count is now
-  // asserted. If the layout default ever changes, this fails loudly instead of
-  // going quietly green.
+  // This carried a layout-mode `test.skip` guard for the three-panel layout,
+  // which renders both panels side by side with no tab buttons (see
+  // `switchToAnnotationsTab` in helpers.ts). MEASURED 2026-09-06 with `-g "panel
+  // switcher"` on the reserved ports: the test reported `passed`, not `skipped`,
+  // so the guard never fired — dead weight of the #1529 shape, where a runtime
+  // skip nothing exercises reads exactly like a pass. Asserting the count makes
+  // a layout-default change fail loudly instead of going quietly green.
   await expect(annotations).toHaveCount(1);
 
   await expect(annotations).toBeVisible({ timeout: 5_000 });
